added voxel collision

+++ src/gfx/chunktable.rs
@@ -180,7 +180,7 @@ impl ChunkVaoTable {
         let view = gamestate.cam.get_view();
         let (x, y, z, axis) = selected;
         let (x, y, z) = get_raycast_voxel(x, y, z, dir, axis);
-        let selectedid = gamestate.world.get_block(x as i32, y as i32, z as i32).id;
+        let selectedid = gamestate.world.get_block(x, y, z).id;
         chunkshader.uniform_vec3f("selected", x as f32, y as f32, z as f32);
         chunkshader.uniform_bool("selectedEmpty", selectedid == EMPTY_BLOCK);
         chunkshader.uniform_matrix4f("view", &view);
